#497 Create routes easier, some changes in AppRoute class, move it from macro
